Improve detection of Windows SDK
Existing build process (as documented in Makefile) doesn't work with a
64-bit Windows SDK. Instead of hacking around to make that work, alter
the build system to try to detect the Windows SDK's setenv.cmd script
and query that "properly".

Generate the file Makefile.winsdk which contains six environment
variables determined by running SetEnv.cmd - these are used to generate
an alternate prefix.

The processing falls back to the older assumption of a 32-bit Visual
Studio 2008 if it fails to locate the SDK.