Add MSVC_DETECT to Makefile
Specifying make MSVC_DETECT=0 ... overrides all of the detection logic for
the Microsoft C Compiler and simply assumes that the environment is
correct. This obviously means that building both 32 and 64-bit CHAINS is
not possible.

checkenv script added which outputs a warning if the environment appears
not to include the Windows SDK, tools or the Microsoft C Runtime. The
script also attempts to determine the architecture of cl.exe and so advise
which CHAINS value can be built with MSVC_DETECT=0. No checking logic
added to Makefile, so specifying CHAINS=msvc64 with an x86 compiler and
vice versa will simply build invalid object files.

Updated CHANGES and also added execute bit to findwinsdk.