Fix compilation with older Microsoft CRTs
The Windows Platform SDK includes a release with a CRT which falls between
Visual Studio .NET 2003 and Visual Studio 2005 and which does not include
the secure versions of many CRT functions, including `_putenv_s`.

In `flexdll_dlopen`, the logic for selecting `setenv`, `putenv` or
`_putenv_s` now correctly relies on the CYGWIN and __STDC_SECURE_LIB__
defines only. The only code change to actual code is the need to declare a
variable at the start of the block to comply with C89 for the Microsoft C
Compiler.