Use /EMITVOLATILEMETADATA:NO with Microsoft Linker
Recent Visual Studio 2019 and 2022 are enabling volatile metadata on
x64 by default. There don't appear to be many technical details
available, but the purpose is to improve emulation of AMD64 on ARM64.
Sometimes the .voltbl sections are copying correctly and then seemingly
minor code changes can break it, causing an obscure and undocumented
error from the Microsoft Linker.

OCaml is starting to add the flag to the compiler to disable its
generation, this patch also stops the linker from attempting to use the
section if it's there.