Package the source package as flexdll
The flexdll package is intended for use when building OCaml in
"bootstrapped" flexdll mode and installs the required source files to
the share folder in an opam switch.

flexdll.opam is now the primary source for the version number for the
package, and both version.ml and version.rc are compiled using the
number taken from there.

Additionally, version.rc is corrected to adopt Windows version format,
so the current release is rendered as 0.42.0.0 instead of 0.0.0.42.