Lazily detect cygpath
Calling cygpath can be slow. If neither -cygpath nor -nocygpath is
specified on the command line, then flexlink used to detect whether
cygpath is present, adding an either unnecessary or at best additional
call to cygpath.

Refactored the logic for cygpath use such that the detection forms part
of the first actual call to cygpath.