Include default libraries in defined_in
The defined_in function is used when determining which objects from the
command line will actually be linked. However, this doesn't include the
default libraries, which creates an issue for modules which are linked
as a result of main.

defined_in is augmented to include the mapping of symbols found in the
default libraries, but some additional care is required to ensure that
this doesn't cause imports from those to leak into the export table of
the executable.