Consider the entrypoint when determining exports
The export table is determined by looping over the objects included on
the command line. If the main symbol is included in a library (e.g.
libcamlrun or libasmrun) then there may not be an object on the
commandline which causes it, and any transitive dependencies, to be
linked.

The entrypoint symbol for the Cygwin/mingw-w64/MSVC toolchains is now
resolved, and may cause additional objects to be linked.