Remove unused intnat/uintnat definitions
Their names are confusing with respect to the OCaml runtime definition
of intnat and uintnat, which are wide enough to store a pointer.
On 64-bits, FlexDLL's uintnat remains a long, thus 32 bits wide.
In all other places the truncation of the pointer is done by casting
to a UINT32*, do that instead.